co_table {vcd}	R Documentation

Compute Conditional Tables

Description

For a contingency table in array form, compute a list of conditional tables given some margins.

Usage

co_table(x, margin, collapse = ".")

Arguments

x
a contingency table in array form.

margin
margin index(es) or corresponding name(s) of the conditioning variables.

collapse
character used when collapsing level names (if more than 1 margin is specified).

Details

This is essentially an interface to [ which is more convenient for arrays of arbitrary dimension.

Value

A list of the resulting conditional tables.

Author(s)

Achim Zeileis Achim.Zeileis@R-project.org

Examples

data("HairEyeColor")
co_table(HairEyeColor, 1)
co_table(HairEyeColor, c("Hair", "Eye"))
co_table(HairEyeColor, 1:2, collapse = "")

[Package vcd version 1.4-7 Index]