mplot {vcd}	R Documentation

Multiple Grid plots

Description

combines severals grid-based plots in a multi-panel-layout.

Usage

mplot(..., .list = list(),
      layout = NULL, cex = NULL,
      main = NULL, gp_main = gpar(fontsize = 20),
      sub = NULL, gp_sub = gpar(fontsize = 15),
      keep_aspect_ratio = TRUE)

Arguments

..., .list
A list of objects inheriting from class "grob", or having a "grob" attribute containing such an object.

layout
integer vector of length 2 giving the number of rows and columns. If NULL, the values will be guessed using some heuristics from the number of objects supplied in ....

cex
Scaling factor for the fonts in the subplots. If NULL, the value is calculated as the inverse square root of the row number.

main, sub
Optional main and sub title, respectively.

gp_main, gp_sub
Optional objects of class "gpar" specifying the graphical parameters for the main and sub title, respectively.

keep_aspect_ratio
logical; should the aspect ratio of the plots be fixed?

Details

This is a convenience function for producing multi-panel plots from grid-based displays, especially those produced by the vcd methods. The layout (number of rows and columns) is guessed from the amount of supplied objects, if not supplied. Currently, the vcd plotting functions do not return grob objects by default—this might change in the future. Also, some of them will return the grob object as a "grob" attribute, attached to the currently returned object.

Value

None.

Author(s)

David Meyer David.Meyer@R-project.org

Examples

mplot(mosaic(Titanic, return_grob = TRUE),
      assoc(Titanic), return_grob = TRUE)

A = mosaic(Titanic, return_grob = TRUE)
B = mosaic(Titanic, type = "expected", return_grob = TRUE)
mplot(A, B)

mplot(sieve(SexualFun, return_grob = TRUE),
      agreementplot(SexualFun, return_grob = TRUE),
      main = "Sexual Fun")

mplot(A, grid.circle())

[Package vcd version 1.4-7 Index]