plot.loglm {vcd}	R Documentation

Visualize Fitted Log-linear Models

Description

Visualize fitted "loglm" objects by mosaic or association plots.

Usage

## S3 method for class 'loglm'
plot(x, panel = mosaic, type = c("observed", "expected"),
  residuals_type = c("pearson", "deviance"), gp = shading_hcl, gp_args = list(),
  ...)

Arguments

x
a fitted "loglm" object, see loglm.

panel
a panel function for visualizing the observed values, residuals and expected values. Currently, mosaic and assoc in vcd.

type
a character string indicating whether the observed or the expected values of the table should be visualized.

residuals_type
a character string indicating the type of residuals to be computed.

gp
object of class "gpar", shading function or a corresponding generating function (see details and shadings). Ignored if shade = FALSE.

gp_args
list of arguments for the shading-generating function, if specified.

...
Other arguments passed to the panel function.

Details

The plot method for "loglm" objects by default visualizes the model using a mosaic plot (can be changed to an association plot by setting panel = assoc) with a shading based on the residuals of this model. The legend also reports the corresponding p value of the associated goodness-of-fit test. The mosaic and assoc methods are simple convenience interfaces to this plot method, setting the panel argument accordingly.

Value

The "structable" visualized is returned invisibly.

Author(s)

Achim Zeileis Achim.Zeileis@R-project.org

See Also

loglm, assoc, mosaic, strucplot

Examples

library(MASS)
## mosaic display for PreSex model
data("PreSex")
fm <- loglm(~ PremaritalSex * ExtramaritalSex * (Gender + MaritalStatus),
  data = aperm(PreSex, c(3, 2, 4, 1)))
fm
## visualize Pearson statistic
plot(fm, split_vertical = TRUE)
## visualize LR statistic
plot(fm, split_vertical = TRUE, residuals_type = "deviance")

## conditional independence in UCB admissions data
data("UCBAdmissions")
fm <- loglm(~ Dept * (Gender + Admit), data = aperm(UCBAdmissions))

## use mosaic display
plot(fm, labeling_args = list(abbreviate = c(Admit = 3)))

## and association plot
plot(fm, panel = assoc)
assoc(fm)

[Package vcd version 1.4-7 Index]