table2d_summary {vcd}	R Documentation

Summary of a 2-way Table

Description

Prints a 2-way contingency table along with percentages, marginal, and conditional distributions.

Usage

table2d_summary(object, margins = TRUE, percentages = FALSE,
  conditionals = c("none", "row", "column"), chisq.test = TRUE, ...)

Arguments

object
a r x c-contingency table

margins
if TRUE, marginal distributions are computed.

percentages
if TRUE, relative frequencies are computed.

conditionals
if not "none", the conditional distributions, given the row/column factor, are computed.

chisq.test
if TRUE, a chi-squared test of independence is carried out.

...
currently not used.

Value

Returns invisibly a r x c x k table, k depending on the amount of choices (at most 3).

Author(s)

David Meyer David.Meyer@R-project.org

See Also

mar_table, prop.table, independence_table

Examples

data("UCBAdmissions")
table2d_summary(margin.table(UCBAdmissions, 1:2))

[Package vcd version 1.4-7 Index]