woolf_test {vcd}	R Documentation

Woolf Test

Description

Test for homogeneity on 2 x 2 x k tables over strata (i.e., whether the log odds ratios are the same in all strata).

Usage

woolf_test(x)

Arguments

x
A 2 x 2 x k table.

Value

A list of class "htest" containing the following components:

statistic
the chi-squared test statistic.

parameter
degrees of freedom of the approximate chi-squared distribution of the test statistic.

p.value
p-value for the test.

method
a character string indicating the type of test performed.

data.name
a character string giving the name(s) of the data.

observed
the observed counts.

expected
the expected counts under the null hypothesis.

References

Woolf, B. 1955. On estimating the relation between blood group and disease. Ann. Human Genet. (London) 19, 251-253.

See Also

mantelhaen.test

Examples

data("CoalMiners")
woolf_test(CoalMiners)

[Package vcd version 1.4-7 Index]